+++ bsn1/CHANGELOG.md
@@ -397,6 +397,7 @@ Nothing is changed.
 - Create method `Error::into\_anyhow()`
 - Create method `Error::root\_cause()`
 - Create method `Error::context()`
+- Create method `Error::as_anyhow()`
 
 ### Removed